edit exception files
