--- src/main/kotlin/domain/utils/TaskExceptions.kt
@@ -1,12 +1,0 @@
-package domain.utils
-
-open class TaskExceptions(message: String) : PlanMateExceptions(message) {
-    class TaskNotFoundException(error: String? = "This task not found") : TaskExceptions(error ?: "This task not found")
-    class TaskCannotDeleteException(error: String = "Can't delete this task") : TaskExceptions(error)
-    class TaskTitleIsEmptyException(error: String = "Task title is empty") : TaskExceptions(error)
-    class TaskDescriptionIsEmptyException(error: String = "Task description is empty") : TaskExceptions(error)
-    class TaskStateIsEmptyException(error: String = "Task State is empty") : TaskExceptions(error)
-    class TaskCannotEditException(error: String = "Can't edit this task") : TaskExceptions(error)
-    class TaskEditFailedException(error: String = "Failed to edit task") : TaskExceptions(error)
-    class NoLogsFoundException(error: String = "There are no logs found for this task") : TaskExceptions(error)
-}
